RT-114: Add load more in route hisotry

+++ Web/ResellTicketSolution/WebClient.Customer/index.js
@@ -16,9 +16,11 @@ function indexController() {
     const model = {
         page: 1,
         pageSize: 10,
+        total: 0,
         routes: [],
         searchStatus: routeStatus.New,
         routeComponents: [],
+        isLoadingMore: false,
     }
 
     if (authenticationService.isLogedin()) {
@@ -41,7 +43,7 @@ function indexController() {
         container.appendChild(routeFilterStatusComponent.render());
     }
 
-    async function renderRoutes() {
+    async function renderRoutes(doNotRemoveChildren) {
         const params = {
             page: model.page,
             pageSize: model.pageSize,
@@ -50,7 +52,10 @@ function indexController() {
 
         const response = await apiService.get(appConfig.apiUrl.route + 'data-table', params);
         const containerElement = document.getElementById(id.routeContainer);
-        commonService.removeAllChildren(containerElement);
+        
+        if(!doNotRemoveChildren) {
+            commonService.removeAllChildren(containerElement);
+        }
 
         response.data.forEach(route => {
             const routeComponent = new RouteComponent(route, onRouteClicked);
@@ -59,6 +64,8 @@ function indexController() {
             routeComponent.render();
             containerElement.appendChild(routeComponent.domElement);
         });
+        model.total = response.total;
+        model.isLoadingMore = false;
     }
 
     function bindEvent() {
@@ -66,6 +73,17 @@ function indexController() {
         btnSearchRoute.addEventListener('click', function() {
             window.location.href = appConfig.url.route.searchForm;
         });
+
+        $(window).scroll(function() {
+            if($(window).scrollTop() + $(window).height() >= ($(document).height() - $('.footer').height())) {
+                if(model.page * model.pageSize < model.total && model.isLoadingMore == false) {
+                    model.isLoadingMore = true;
+
+                    model.page++;
+                    renderRoutes(true);
+                } 
+            }
+        });  
     }
 
     function onRouteClicked(route) {
